Simplify dependencies between *-types packages.
no <service>-types package:
- depends on any other <service>-types package.
- conditionally (cabal flag) depends on cql.
