+++ libs/galley-types/galley-types.cabal
@@ -10,10 +10,6 @@ build-type:    Simple
 cabal-version: >= 1.10
 description:   API types of Galley.
 
-flag cql
-    description: Enable cql instances
-    default:     False
-
 library
     default-language: Haskell2010
     default-extensions: NoImplicitPrelude
@@ -31,10 +27,6 @@ library
         Galley.Types.Teams.Internal
         Galley.Types.Teams.Intra
         Galley.Types.Teams.Swagger
-
-    if flag(cql)
-        cpp-options:   "-DWITH_CQL"
-        build-depends: cql >= 3.0
 
     build-depends:
         aeson                 >= 0.6
