+++ libs/types-common/types-common.cabal
@@ -40,6 +40,7 @@ library
           Data.List1
           Data.Misc
           Data.Range
+          Data.Role
           Data.Swagger
           Data.Text.Ascii
           Data.UUID.Tagged
