+++ services/galley/test/integration/API/MessageTimer.hs
@@ -9,6 +9,7 @@ import Bilge hiding (timeout)
 import Bilge.Assert
 import Data.List1
 import Data.Misc
+import Data.Role
 import Galley.Types
 import Network.Wai.Utilities.Error
 import Test.Tasty
@@ -84,7 +85,7 @@ messageTimerChangeGuest g b _ca _ = do
     -- Create a team and a guest user
     [owner, member, guest] <- randomUsers b 3
     connectUsers b owner (list1 member [guest])
-    tid <- createTeam g "team" owner [Teams.newTeamMember member Teams.fullPermissions Nothing]
+    tid <- createTeam g "team" owner [Teams.newTeamMember member fullPermissions Nothing]
     -- Create a conversation
     cid <- createTeamConv g owner tid [member, guest] Nothing Nothing Nothing
     -- Try to change the timer (as the guest user) and observe failure
