+++ stack.yaml
@@ -169,9 +169,6 @@ flags:
     protobuf: True
     arbitrary: True
 
-  galley-types:
-    cql: True
-
 extra-package-dbs: []
 
 allow-newer: False
